IO_LOTTERY/pyproject.toml
[tool.poetry]
name = "IO_LOTTERY"
version = "0.1.0"
description = "Flask app"
authors = ["Kacper Żebrowski"]
readme = "README.md"
packages = [{include = "IO_LOTTERY"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
